+++ WinForms/MainView.cs
@@ -1,4 +1,4 @@
-namespace InputTracker
+namespace nardnob.InputTracker.WinForms
 {
     public partial class MainView : Form
     {
